+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 - Fix special case issue with keybinds and refactor key handling..
 - Shorten french translations
 - Small improvements overall and allow holding the zoom in/out keys
-- Safer sensitivity restore on leaving
+- Safer/better sensitivity restore on leaving
+- Fix issues with FOV
 
 ### Version 1.20.1-1.20.5 - 2.4.1 (Forge) 08/14/2025
 - Prevent zooming and opening the configuration GUI when being in inventory/gui screens